+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed an issue with the ECMWF platform not working with the new online recovery
 - Changed logic in Slurm submission to look for submitted messaged (hotfix) #2886" 
 - Calendar splits: last split boundaries #2866, last split not exported #2099 and splitsize not working at experiment level #2016.
+- Accept local/intranet email addresses in validation #1471
 
 **Enhancements:**